pyproject.toml
[tool.poetry]
name = "fastapi-jwt"
version = "0.1.0"
description = ""
authors = ["Bruno Fioreze"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
PyJWT = "^2.4.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
